// src/modules/comercial/produtos/roles/fornecedor/dtos/produto-importacao.dto.ts
export default interface ProdutoImportacaoDTO {
  codigo_produto_fornecedor: string;
  descritivo_pdv: string;
  descritivo: string;
  origem: string;
  estado: string;
  preco: string;
  desconto_p: string;
  eans: string;
  duns: string;
  pesob: string;
  pesol: string;
  altura: string;
  largura: string;
  comprimento: string;
  validade: string;
  qtde_embalagem: string;
  comprimento_d: string;
  largura_d: string;
  altura_d: string;
  classificacao_fiscal: string;
  st_compra: string;
  icms_compra: string;
  ipi: string;
  pis_cofins: string;
  descricao: string;
  caracteristica: string;
  modo_uso: string;
  imagens: string;
}